requirements.txt
annotated-types==0.6.0
anyio==3.7.0
blinker==1.6.2
certifi==2023.5.7
charset-normalizer==3.2.0
click==8.1.3
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.1.1
fastapi==0.111.0
fastapi-cli==0.0.2
Flask==2.3.2
future==0.18.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.4
iniconfig==2.0.0
iso8601==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
markdown-it-py==3.0.0
MarkupSafe==2.1.3
mdurl==0.1.2
numpy==1.24.3
orjson==3.10.3
packaging==24.1
pandas==2.0.2
pluggy==1.5.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
pyserial==3.5
pytest==8.3.2
pytest-asyncio==0.24.0
pytest-mock==3.14.0
python-dateutil==2.8.2
python-dotenv==1.0.1
python-multipart==0.0.9
pytz==2023.3
PyYAML==6.0
requests==2.31.0
rich==13.7.1
shellingham==1.5.4
six==1.16.0
sniffio==1.3.0
starlette==0.37.2
tomli==2.0.1
typer==0.12.3
typing_extensions==4.11.0
tzdata==2023.3
ujson==5.9.0
urllib3==2.0.3
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
Werkzeug==2.3.4
